brush up logs and refactoring
