+++ Sources/RIBsCodeGen/Extensions/Array+Extension.swift
@@ -1,0 +1,18 @@
+//
+// Created by 今入　庸介 on 2022/05/11.
+//
+
+import Foundation
+extension Array {
+    func prettyPrinted() {
+        guard let jsonData = try? JSONSerialization.data(withJSONObject: self, options: [.prettyPrinted]) else {
+            print("JSON Serialization error")
+            return
+        }
+        guard let jsonString = String(data: jsonData, encoding: .utf8) else {
+            print("JSON Encoding error")
+            return
+        }
+        print(jsonString)
+    }
+}
